# oci-docs-Realese_IB_0125/docs/swagger.yaml
definitions:
  model.AccountReportCategory:
    properties:
      amount:
        type: number
      nameKz:
        type: string
      nameRu:
        type: string
      number:
        type: string
    type: object
  model.AccountReportsResult:
    properties:
      categories:
        description: UnionType      AccountReportsUnionType `json:"unionType"`
        items:
          $ref: '#/definitions/model.AccountReportCategory'
        type: array
      osiAddress:
        type: string
      osiName:
        type: string
      period:
        description: '"2023-11-30T09:34:03.629Z",'
        type: string
      signer:
        type: string
      unionTypeKz:
        type: string
      unionTypeRu:
        type: string
    type: object
...
  model.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      result:
        type: string
    type: object
...
  model.ReportResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      result:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/report/abonents/{id}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Получить таблицу помещений.
      operationId: GetOSIAbonentsReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получить таблицу помещений.
      tags:
      - Отчеты
  /api/report/accounts/monthly:
    post:
      consumes:
      - application/json
      description: Формирование ежемесячного отчета по счетам.
      operationId: GetAccountsMonthlyReport
      parameters:
      - description: язык отчета
        in: query
        name: language
        type: string
      - description: модель отчета
        in: body
        name: Model
        required: true
        schema:
          $ref: '#/definitions/model.AccountReportsResult'
      produces:
      - application/json
      responses:
        "200":
          description: File Base64
          schema:
            $ref: '#/definitions/model.ReportResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Формирование ежемесячного отчета по счетам.
      tags:
      - Отчеты
  /api/report/debts/{id}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Получение отчета  по долгам за текущий месяц.
      operationId: GetDebtsReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      - description: дата начала
        in: query
        name: begin
        required: true
        type: string
      - description: дата окончания
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получение отчета по долгам за текущий месяц.
      tags:
      - Отчеты
  /api/report/fixes/{id}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Корректировки по всем абонентам и услугам ОСИ за период.
      operationId: GetFixesReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      - description: дата начала
        in: query
        name: begin
        required: true
        type: string
      - description: дата окончания
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Корректировки по всем абонентам и услугам ОСИ за период.
      tags:
      - Отчеты
  /api/report/osv/{id}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Получение отчета сальдо за текущий месяц.
      operationId: GetOSVReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      - description: дата начала
        in: query
        name: begin
        required: true
        type: string
      - description: дата окончания
        in: query
        name: end
        required: true
        type: string
      - description: для абонента
        in: query
        name: forAbonent
        type: boolean
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получение отчета сальдо за текущий месяц.
      tags:
      - Отчеты
  /api/report/osv/{id}/abonent/{abonentId}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Получение отчета сальдо по абоненту за все периоды.
      operationId: GetAbonentOSVReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      - description: id Абонента
        in: path
        name: abonentId
        required: true
        type: integer
      - description: Номер помещения
        in: query
        name: flat
        required: true
        type: string
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получение отчета сальдо по абоненту за все периоды.
      tags:
      - Отчеты
  /api/report/payments/{id}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Получение реестра платежей за период.
      operationId: GetPaymentsReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      - description: дата начала
        in: query
        name: begin
        required: true
        type: string
      - description: дата окончания
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Получение реестра платежей за период.
      tags:
      - Отчеты
  /api/report/payments/orders/{id}:
    get:
      consumes:
      - application/json
      - application/octet-stream
      description: Свод платежных поручений по ОСИ за период.
      operationId: GetPaymentOrdersReport
      parameters:
      - description: id OSI
        in: path
        name: id
        required: true
        type: integer
      - description: дата начала
        in: query
        name: begin
        required: true
        type: string
      - description: дата окончания
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Свод платежных поручений по ОСИ за период.
      tags:
      - Отчеты
...
swagger: "2.0"